slack component restructure
