+++ gateway/tests/slack/test_channel_intro.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from gateway.transports.slack.channel_intro import ChannelIntroGreeter
+from gateway.transports.slack.outbound.channel_intro import ChannelIntroGreeter
 
 
 class _FakeMessagingClient:
