+++ gateway/tests/slack/test_client.py
