+++ gateway/tests/slack/test_dispatcher_scope_borders.py
@@ -23,12 +23,12 @@ from config.constants.billing import ORGANIZATION_ID_ENV, USAGE_SECRET_ENV, WEBA
 from config.principal import Actor, Principal
 from config.scope_context import current_scope
 from gateway.core.billing.credits_client import CreditsOutcome
-from gateway.transports.slack.dispatcher import _SlackTurnDispatcher
-from gateway.transports.slack.events import SlackInboundMessage
-from gateway.transports.slack.principal import slack_scope
+from gateway.transports.slack.inbound.dispatcher import _SlackTurnDispatcher
+from gateway.transports.slack.inbound.events import SlackInboundMessage
+from gateway.transports.slack.inbound.principal import slack_scope
 from gateway.transports.slack.settings import SlackGatewaySettings
 
-_SECURITY = "gateway.transports.slack.security"
+_SECURITY = "gateway.transports.slack.inbound.security"
 TEST_ORG = "org_border_acme"
 
 
@@ -214,7 +214,7 @@ def test_consume_credits_uses_org_principal_not_slack_user(
         billed.append(organization_id)
         return CreditsOutcome.UNCONFIGURED
 
-    monkeypatch.setattr("gateway.transports.slack.dispatcher.consume_credits", _consume)
+    monkeypatch.setattr("gateway.transports.slack.inbound.dispatcher.consume_credits", _consume)
     _dispatcher(resolver=_RecordingResolver(), handler=lambda *_a: None).dispatch(
         _inbound(user_id="U_ALICE")
     )
