+++ gateway/tests/slack/test_settings.py
