+++ gateway/tests/slack/test_turn_stream.py
@@ -1,4 +1,4 @@
-"""TDD: Slack ``_TurnStream`` answer buffering must not use string ``+=``.
+"""TDD: Slack ``TurnStream`` answer buffering must not use string ``+=``.
 
 Many tiny markdown chunks arrive on the stream path. Buffering with
 ``s += chunk`` is quadratic; a part list + ``join`` on flush is linear.
@@ -9,15 +9,15 @@ from __future__ import annotations
 
 from unittest.mock import MagicMock
 
-from gateway.transports.slack.output_sink import _TurnStream
+from gateway.transports.slack.outbound.turn_stream import TurnStream
 
 
-def _stream(*, update_interval_seconds: float = 10_000.0) -> tuple[_TurnStream, MagicMock]:
+def _stream(*, update_interval_seconds: float = 10_000.0) -> tuple[TurnStream, MagicMock]:
     client = MagicMock()
     client.start_stream.return_value = "ts.1"
     client.append_stream.return_value = True
     client.stop_stream.return_value = True
-    stream = _TurnStream(
+    stream = TurnStream(
         client=client,
         channel_id="C1",
         thread_ts="t.1",
