+++ gateway/transports/slack/connection/__init__.py
@@ -1,0 +1,1 @@
+"""Slack Socket Mode connection: worker lifecycle and liveness heartbeat."""
