+++ gateway/transports/slack/inbound/__init__.py
@@ -1,0 +1,1 @@
+"""Inbound Slack turns: payload parsing, authorization, and dispatch."""
