+++ gateway/transports/slack/outbound/__init__.py
@@ -1,0 +1,1 @@
+"""Outbound Slack delivery: streamed replies, approvals, and feedback."""
